demand parameter from config
Former-commit-id: c896b337018b85bc6c1fc63f53342380e44c0c57 [formerly 2869b1d9069d75c8737a7fdb0f317b2dafe33e03]
Former-commit-id: d81b09ea241c910b38b778a8973196583fbd3764

+++ AdX/src/main/java/tau/tac/adx/demand/CampaignImpl.java
@@ -49,7 +49,7 @@ public class CampaignImpl implements Campaign, Accumulator<CampaignStats> {
 	
 	
 	
-	public CampaignImpl(QualityManager qualityManager, Long reachImps, int dayStart, int dayEnd,
+	public CampaignImpl(QualityManager qualityManager, int reachImps, int dayStart, int dayEnd,
 			MarketSegment targetSegment, double videoCoef, double mobileCoef) {
 		
 		if (qualityManager == null)
@@ -64,7 +64,7 @@ public class CampaignImpl implements Campaign, Accumulator<CampaignStats> {
 		day = 0;
 		
 		this.qualityManager = qualityManager;
-		this.reachImps = reachImps;
+		this.reachImps = (long)reachImps;
 		this.dayStart = dayStart;
 		this.dayEnd = dayEnd;
 		this.targetSegment = targetSegment;
